added struct encapsulation

+++ cpp/lab3/qns10.cpp
@@ -1,9 +1,11 @@
 // Write a recursive c++ program to find the factorial of entered number
 
 #include <iostream>
+
+
 using namespace std;
 
-int factorial(int n)
+long int factorial(long int n)
 {
     if (n == 0)
         return 1;
